first

+++ Area of circle.c
@@ -2,7 +2,7 @@
 #define pi 3.1415
 int main()
 {
-    int r = 5;
+    float r = 5;
     float area, circum;
     area = pi * r * r;
     circum = 2 * pi * r;
